Break up project definitions
