+++ project/src/main/scala/modules/allDeps.scala
@@ -1,0 +1,23 @@
+package dev.guardrail.sbt.modules
+
+import sbt._
+import sbt.Keys._
+
+import dev.guardrail.sbt.Build._
+import dev.guardrail.sbt.Dependencies._
+
+object allDeps {
+  val project = Project("allDeps", file("modules/alldeps"))
+    .settings(commonSettings)
+    .settings(
+      publish / skip := true,
+      libraryDependencies ++= scalaAkkaHttp.dependencies,
+      libraryDependencies ++= scalaAkkaHttp.dependenciesJackson,
+      libraryDependencies ++= scalaHttp4s.dependencies,
+      libraryDependencies ++= scalaEndpoints.dependencies,
+      libraryDependencies ++= javaSpringMvc.dependencies,
+      libraryDependencies ++= javaDropwizard.dependencies,
+      libraryDependencies ++= javaDropwizard.dependenciesVavr,
+      libraryDependencies ++= scalaDropwizard.dependencies,
+    )
+}
