+++ project/src/main/scala/modules/guardrail.scala
@@ -1,0 +1,10 @@
+package dev.guardrail.sbt.modules
+
+import dev.guardrail.sbt.Build._
+
+import sbt._
+import sbt.Keys._
+
+object guardrail {
+  val project = baseModule("guardrail", "guardrail", file("modules/codegen"))
+}
