+++ project/src/main/scala/modules/root.scala
@@ -1,0 +1,13 @@
+package dev.guardrail.sbt.modules
+
+import sbt._
+import sbt.Keys._
+
+import dev.guardrail.sbt.Build._
+
+object root {
+  val project = Project("root", file("."))
+    .settings(commonSettings)
+    .settings(publish / skip := true)
+    .settings(libraryDependencies += "org.slf4j" % "slf4j-simple" % "1.7.32")
+}
